+++ static/app/gettingStartedDocs/node/awslambda/awslambdaArnSelector.tsx
