+++ static/app/gettingStartedDocs/node/awslambda/onboarding.tsx
@@ -2,8 +2,8 @@ import {ExternalLink} from 'sentry/components/core/link';
 import {StepType} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import type {OnboardingConfig} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {getUploadSourceMapsStep} from 'sentry/components/onboarding/gettingStartedDoc/utils';
+import {AwsLambdaArn} from 'sentry/gettingStartedDocs/node/awslambda/awslambdaArnSelector';
 import {InstallationMethod} from 'sentry/gettingStartedDocs/node/awslambda/utils';
-import {AwsLambdaArn} from 'sentry/gettingStartedDocs/node/awslambdaArnSelector';
 import {t, tct} from 'sentry/locale';
 import {getInstallCodeBlock} from 'sentry/utils/gettingStartedDocs/node';
